Remimplement json encoding of inlines manually.
This is the first step to doing manual encoding and decoding of pandoc
JSON. This will replace the current generic deriving, which can be a
moving target. It should also allow us to clean up the JSON format in
the future. Note that in this first step, we are only seeking to
reimplement the current format.