Abstract out encoding function for Pandoc elements
Instead of explicitly encoding every "t"/"c" pair, use `tagged` and
`taggeNoContent` functions. Note that these have to be defined
separately because of type variable in `tagged`.

This also abbreviates encoding of arrays by taking advantage of the
ToJSON instances of tuples (up to size 7). Note that if we have any
elements with more than 7 arguments, we'll have to explicitly encode an
array. (We also remove the OverloadedLists extension, because we don't
need it if we use tuples.)