Add explicit unit tests for encoding and decoding.
Previously we had just done round-trip tests, but that can hide a
mistake if the encoder and decoder are making the same mistake (or if
the encoder makes a mistake that the decoder ignores). This tests each
pandoc entity against a specific JSON string. There is some redundancy
because of the recursive nature of pandoc types, but testing each type
individually will help us pinpoint problems.

Note that this uses a simple quasiquoting library (strings-qq) to avoid
escaping every quote in a JSON string.