Provide default implementation for walk
The implementations for `walk` and `walkM` are very similar, so a
default method is provided which implements the former in terms of the
latter. This change should not affect performance, as the `Identity`
functor, which is used in the default definition, is a newtype that
should be eliminated at compile time.