Force optimizations when compiling Walk module
The functions defined in the Walk module are relatively simple while
being called recursively on possibly long documents. It hence seems
sensible to force optimizations for this module, as the benefits likely
outweigh the additional compile time.

This also reduces the performance cost of having a default
implementation for `walk` based on `walkM`.