Text.Pandoc.JSON: Use walk instead of bottomUp for a -> [a].
Note that behavior will be slightly different, since
bottomUp's treatment of a function f : [a] -> [a] is to
apply it to each sublist of a list, while walk applies
it only to maximal sublists.

Usually the latter behavior is what is wanted, and the
former can be simulated when needed.