Add Text.Pandoc.Legacy.Definition
To ease the transition to Text, this module provides an interface
compatible with the String one, so that any unqualified imports of
Text.Pandoc.Definition in other packages can be replaced by
Text.Pandoc.Legacy.Definition without other code changes. This is done
with PatternSynonyms.

Some of the constructors of the types Meta, MetaValue, Block, Inline,
Format, and Citation required PatternSynonym handling. The Attr and
Target types had to be redefined, and certain functions had to be
rewritten to handle String or the old Attr and Target types in this
module. This module otherwise exports the definitions in
Text.Pandoc.Definition unchanged.

This is not a perfect drop-in replacement, since some imports like
Inline(..) will no longer work. This may also cause incomplete pattern
warnings when used, since the coverage checker does not seem to be
aware of PatternSynonyms.