Fix tests, small caption bug
The caption can now contain block content. Now, if a non-empty inline
caption is given, it is wrapped in a Para block, and if it is empty
then a fully empty caption should be added to the table. Before, the
wrapping would happen unconditionally.

This is only temporary. When the builder signatures change this will
become unnecessary.