Use untagged JSON encoding for single-constructor types
All of the single constructor types related to Table are now
represented in JSON either as arrays (for multi-argument constructors)
or as the representation of the inner type (for single argument
constructors). This behaviour for newtype-defined and multi-argument
non-record types is now consistent across the entire JSON interface,
with the exception of Pandoc itself (which is represented as a JSON
object with additional metadata). Multi-argument records (of which
Citation is the only example) are still represented as objects with
the record accessors as keys.