Change the table builder to permit looser intermediate table heads
The table builder (and the normalizeTableBody function) now permit
cells in the intermediate head of a TableBody to extend past the
RowHeadColumns. This allows for intermediate tables to have
subheadings that extend across the entire table.

Formerly the table builder would treat the intermediate head like the
intermediate body, and clip or drop cells that extended past the row
head.