Add CustomFormat, modify JATS sub-format relation
The new CustomFormat will allow for filter writers to use raw elements
with custom formats, and to use custom formats in
conditionally-rendered elements. They are invisible to castsToKnown
and knownMatch.

The Formats type had to be modified to accommodate that change, and
additional conversion and convenience functions were added.

jats_articleauthoring is a super-format of jats_publishing. The
listSubformats relation was modified to take this into account.