[API change] Modify Attr, add it to relevant types
The new Attr type, in addition to being a data type, changes the
key-value pairs representation to a Map.

An Attr component has been added to every relevant branch of Inline
and Block. It has also been added to Caption.

New Attr conversion functions have been added to Definition, and new
builders for the types that received an Attr component have been added
to Builder.

The representation of Attr in JSON is different: the key-value pairs
are now an object.

Additionally, the t_table test was a bit too large to inspect for
correctness directly. It has been simplified, and new tests for the
top-level components of Table have been added.