[API change] Add a Ref inline element and Reference type
The Ref element is analogous to the \ref family of LaTeX commands. It
is a collection of references to other identified elements, similar to
the Cite element. It includes a RefType type that distinguishes
between number references and page references.

The Reference type represents a single reference to an identified
element. Unlike a Link, the text of a Reference is generally expected
to be replaced with attributes of the referenced element. These
attributes are normally either its number or page number, and a
possibly-configurable prefix representing the type of thing referenced
(like "figure" or "table").

The ReferenceMode type records any local modification to the prefix of
a reference, recognizing explicit upper case and lower case prefix
selection, prefix suppression, and normal (default) references.