Use untagged JSON encoding for single-constructor types (#76)
All of the single constructor types related to Table are now
represented in JSON either as arrays (for multi-argument constructors)
or as the representation of the inner type (for single argument
constructors). This behaviour for newtype-defined and multi-argument
non-record types is now consistent across the entire JSON interface,
with the exception of Pandoc itself (which is represented as a JSON
object with additional metadata). Multi-argument records (of which
Citation is the only example) are still represented as objects with
the record accessors as keys.

The Meta and Citation types now use derived JSON serialization
(newtype and generic, respectively). The format remains the same as
before.

New serialization tests now test that Meta and the Table types are
encoded properly in JSON. The Citation type already had such a test.

Use TH To/FromJSON instances